+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/ProjectileSpell.java
@@ -1,7 +1,10 @@
 package com.elmakers.mine.bukkit.plugins.magic.spells;
 
+import org.bukkit.Location;
+import org.bukkit.Sound;
 import org.bukkit.entity.Fireball;
 import org.bukkit.entity.Projectile;
+import org.bukkit.entity.WitherSkull;
 
 import com.elmakers.mine.bukkit.plugins.magic.Spell;
 import com.elmakers.mine.bukkit.plugins.magic.SpellResult;
@@ -20,13 +23,20 @@ public class ProjectileSpell extends Spell
 		boolean useFire = parameters.getBoolean("fire", true);
 		String projectileClass = parameters.getString("projectile", "Fireball");
 		try {
+			Location playerLocation = player.getLocation();
 			Class<? extends Projectile> projectileType = (Class<? extends Projectile>)Class.forName("org.bukkit.entity." + projectileClass);
 			Projectile projectile = player.launchProjectile(projectileType);
 			projectile.setShooter(player);
+			if (projectile instanceof WitherSkull) {
+				playerLocation.getWorld().playSound(playerLocation, Sound.WITHER_SHOOT, 1.0f, 1.5f);		
+			}
 			if (projectile instanceof Fireball) {
 				Fireball fireball = (Fireball)projectile;
 				fireball.setIsIncendiary(useFire);
 				fireball.setYield(size);
+				if (!(projectile instanceof WitherSkull)) {
+					playerLocation.getWorld().playSound(playerLocation, Sound.GHAST_FIREBALL, 1.0f, 1.5f);
+				}
 			}
 		} catch(Exception ex) {
 			sendMessage("Unknown projectile class " + projectileClass);